discord_bot/requirements.txt
discord.py>=2.3.0
httpx>=0.27.0
anthropic>=0.30.0
python-dotenv>=1.0.0
